Day 9 AoC the Loooooooong version

+++ src/main/kotlin/com/sanderverbruggen/adventofcode/day7/AmplifierRunner.kt
@@ -7,23 +7,23 @@ import kotlinx.coroutines.runBlocking
 
 class AmplifierRunner(private val intCode: String) {
 
-    fun solve(): Int {
+    fun solve(): Long {
         return permute(listOf(0, 1, 2, 3, 4))
                 .map { runAmpChain(it) }
                 .max() ?: -1
     }
 
-    fun solveLoopback(): Int {
+    fun solveLoopback(): Long {
         return permute(listOf(5, 6, 7, 8, 9))
                 .map { runAmpChain(it) }
                 .max() ?: -1
     }
 
 
-    internal fun runAmpChain(phases: List<Int>): Int {
+    internal fun runAmpChain(phases: List<Int>): Long {
         return runBlocking {
-            val channels = phases.map { phase -> Channel<Int>(4).apply { send(phase) } }.toMutableList()
-            channels.first().send(0) // initial input
+            val channels = phases.map { phase -> Channel<Long>(4).apply { send(phase.toLong()) } }.toMutableList()
+            channels.first().send(0L) // initial input
             runBlocking {
                 for (index in 0 until phases.size) {
                     launch {
